+++ app/middlewares/navigation/index.js
@@ -1,3 +1,2 @@
 export { default } from './middleware'
-export { default as rootNavigator } from './navigators'
 export { setTopLevelNavigator } from './middleware'
