+++ app/navigation/app-container.js
@@ -1,0 +1,4 @@
+import { createAppContainer } from 'react-navigation'
+import rootNavigator from 'hg/navigation/navigators'
+
+export default createAppContainer(rootNavigator)
