+++ app/navigation/navigators/onboarding.js
@@ -1,5 +1,5 @@
 import { createSwitchNavigator } from 'react-navigation'
-import { ONBOARDING_AGREE } from 'hg/middlewares/navigation/routes'
+import { ONBOARDING_AGREE } from 'hg/navigation/routes'
 import {
   OnboardingAgree,
   OnboardingInitialize,
